requirements.txt
# data science
pandas

# dataviz
plotly.express
dash
dash-bootstrap-components
dash_bootstrap_templates

# GPX
gpxpy
fitdecode

# Server
gunicorn
